feat:迁移API相关数据模型至单独文件定义
